+++ R/ebv_create.R
@@ -1,13 +1,13 @@
 #' Create an EBV netCDF
 #'
-#' @description Create the core structure of the EBV NetCDF based on the json
+#' @description Create the core structure of the EBV netCDF based on the json
 #'   from the \href{https://portal.geobon.org/api-docs}{Geobon Portal API}. Data
 #'   will be added afterwards. Use [ebvcube::ebv_add_data()] to add the missing
 #'   data.
 #'
 #' @param jsonpath Character. Path to the json file downloaded from the
 #'   \href{https://portal.geobon.org/api-docs}{Geobon Portal API}.
-#' @param outputpath Character. Set path where the NetCDF file should be
+#' @param outputpath Character. Set path where the netCDF file should be
 #'   created.
 #' @param entities Character string or vector of character strings. In case of
 #'   single character string: Path to the csv table holding the entity names.
@@ -256,11 +256,18 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
   #get temp directory
   temp_path <- tempdir()
 
-  #check timesteps
-  if(!is.null(timesteps)){
+  #read json ----
+  file <- jsonlite::fromJSON(txt=jsonpath)
+  #json root
+  json <- file$data
+
+  #check timesteps----
+  t_res <- json$time_coverage$time_coverage_resolution
+
+  if(!is.null(timesteps) & t_res != 'Paleo'){
     if (checkmate::checkCharacter(timesteps) != TRUE){
       stop('timesteps needs to be a list of character values.')
-    }else{
+    }else {
       for(ts in timesteps){
         #check ISO format
         if(!(grepl('^\\d{4}-\\d{2}-\\d{2}$', ts) | grepl('^\\d{4}$',ts))){
@@ -285,10 +292,7 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
 
   }
 
-  #read json ----
-  file <- jsonlite::fromJSON(txt=jsonpath)
-  #json root
-  json <- file$data
+
 
   # get basic hierarchy info ----
   metrics_no <- length(json$ebv_metric)
@@ -328,15 +332,41 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
 
   # get dimensions ----
   # time ----
-  t_res <- json$time_coverage$time_coverage_resolution
   t_start <- json$time_coverage$time_coverage_start
   t_end <- json$time_coverage$time_coverage_end
 
-  #create timesteps
+  #get ISO timesteps for irregular and paleo -> shiny app
+  if(t_res=='Irregular'){
+    if(is.null(timesteps)){
+      timesteps <- json$timesteps[[1]]
+    }
+    if(!is.null(timesteps)){
+      if(timesteps[1]=='N/A'){
+        timesteps <- NULL
+      }
+    }
+  }
+
+  if(t_res=='Paleo'){
+    if(is.null(timesteps)){
+      timesteps <- json$timesteps[[1]]
+    }
+    if(!is.null(timesteps)){
+      if(timesteps[1]=='N/A'){
+        timesteps <- NULL
+      }
+    }
+    if(!is.null(timesteps)){
+      timesteps <- as.numeric(timesteps)
+      timesteps <- sort(timesteps, decreasing = TRUE)
+    }
+  }
+
+  #create integer timesteps
   add <- 40177
 
   #calculate timesteps
-  if(is.null(timesteps)){
+  if(is.null(timesteps) & t_res!='Paleo'){
     if(t_res=="P0000-00-00"){
       #one timestep only
       #check
@@ -435,7 +465,7 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
         timesteps <- c(0)
       }
     }
-  }else{
+  }else if (t_res != 'Paleo'){
     #take given timesteps and transform them into integer values
     temp_temp <- c()
     for (ts in timesteps){
@@ -450,7 +480,10 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
     }
     timesteps <- temp_temp
   }
-
+  #if no timesteps are presented anywhere, throw error
+  if(is.null(timesteps)){
+    stop('There are no timesteps given. Define the argument "timesteps", to create your EBV netCDF.')
+  }
 
 
   # lat ----
@@ -480,7 +513,11 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
   # create dimensions ----
   lat_dim <- ncdf4::ncdim_def('lat', crs_unit , vals = lat_data)
   lon_dim <- ncdf4::ncdim_def('lon', crs_unit, vals = lon_data)
-  time_dim <- ncdf4::ncdim_def('time', 'days since 1860-01-01 00:00:00.0' , timesteps, unlim = T)#HERE
+  if(t_res=='Paleo'){
+    time_dim <- ncdf4::ncdim_def('time', 'kyrs B.P.' , timesteps, unlim = T)#HERE
+  }else{
+    time_dim <- ncdf4::ncdim_def('time', 'days since 1860-01-01 00:00:00.0' , timesteps, unlim = T)#HERE
+  }
   entity_dim <- ncdf4::ncdim_def('entity', '', vals = 1:entities_no, create_dimvar=FALSE)
 
   # create list of vars 3D ----
@@ -738,7 +775,7 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
   for (i in 1:length(global.att)){
     att.txt <- eval(parse(text = paste0('json$', global.att[i][[1]])))
     att.txt <- paste0(trimws(att.txt), collapse = ', ')
-    if(names(global.att[i])=='contributor_name'){
+    if(names(global.att[i])=='contributor_name' | names(global.att[i])=='ebv_domain'){
       att.txt <- paste0(trimws(trimws(stringr::str_split(att.txt,',')[[1]])), collapse = ', ')
     }
     ebv_i_char_att(hdf, names(global.att[i]), att.txt)
@@ -926,7 +963,9 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
   ebv_i_char_att(time.id, 'axis', 'T')
 
   # :calendar = "standard";
-  ebv_i_char_att(time.id, 'calendar', 'standard')
+  if(t_res != 'Paleo'){
+    ebv_i_char_att(time.id, 'calendar', 'standard')
+  }
 
   #close
   rhdf5::H5Dclose(time.id)
@@ -1008,8 +1047,8 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
       enum <- as.integer(paste0(stringr::str_extract_all(part, '\\d')[[1]], collapse=''))
       did <- rhdf5::H5Dopen(hdf, var)
       ebv_i_char_att(did, 'grid_mapping', '/crs')
-      ebv_i_char_att(did, 'coordinate', '/entity')#HERE
-      ebv_i_char_att(did, 'coverage_content_type', paste0(json$coverage_content_type[[1]], collapse=', '))
+      ebv_i_char_att(did, 'coordinates', '/entity')#HERE
+      ebv_i_char_att(did, 'coverage_content_type', paste0(json$coverage_content_type, collapse=', '))
       ebv_i_char_att(did, 'standard_name', entity_csv[enum,1])
       #close dh
       rhdf5::H5Dclose(did)
@@ -1027,8 +1066,8 @@ ebv_create <- function(jsonpath, outputpath, entities, epsg = 4326,
       did <- rhdf5::H5Dopen(hdf, var)
       ebv_i_char_att(did, 'long_name', long_name)
       ebv_i_char_att(did, 'grid_mapping', '/crs')
-      ebv_i_char_att(did, 'coordinate', '/entity')#HERE
-      ebv_i_char_att(did, 'coverage_content_type', paste0(json$coverage_content_type[[1]], collapse=', '))
+      ebv_i_char_att(did, 'coordinates', '/entity')#HERE
+      ebv_i_char_att(did, 'coverage_content_type', paste0(json$coverage_content_type, collapse=', '))
       #close dh
       rhdf5::H5Dclose(did)
     }